Adding namespace, removing redundant code, improving comments
