+++ Editor/Unity.VSP-Analytics.Editor.asmdef
@@ -1,5 +1,5 @@
 {
-    "name": "Unity.VSP-Analytics.Runtime",
+    "name": "Unity.VSP-Analytics.Editor",
     "references": [],
     "includePlatforms": [],
     "excludePlatforms": [],
